refactor(module-context): rename context provider file
